Refactor for maintainability
