+++ ProcessFewsEventCode/helpers/get-boolean-indicator.js
@@ -1,7 +1,4 @@
-const extract = require('./extract')
-
-const expectedNumberOfMatches = 2
-const matchIndexToReturn = 1
+const extract = require('./extraction-utils')
 
 module.exports = async function (context, taskRunData, indicatorName) {
   let indicator = taskRunData.message.includes('is made current manually')
@@ -10,15 +7,7 @@ module.exports = async function (context, taskRunData, indicatorName) {
     const indicatorRegex = new RegExp(`(?:${indicatorName}\\:?\\s*(True|False))`, 'i')
     const indicatorText = `task run ${indicatorName} status`
 
-    const extractionData = {
-      taskRunData,
-      regex: indicatorRegex,
-      expectedNumberOfMatches,
-      matchIndexToReturn,
-      errorMessageSubject: indicatorText
-    }
-
-    const indicatorString = await extract(context, extractionData)
+    const indicatorString = await extract(context, taskRunData, indicatorRegex, indicatorText)
 
     if (typeof indicatorString === 'undefined') {
       indicator = undefined
