+++ ProcessFewsEventCode/helpers/get-task-run-id.js
@@ -1,18 +1,8 @@
-const extract = require('./extract')
+const extract = require('./extraction-utils')
+
+const taskRunIdRegex = /\sid(?:\s*=?\s*)([^\s)]*)(?:\s*\)?)/i
+const taskRunIdText = 'task run ID'
 
 module.exports = async function (context, taskRunData) {
-  const expectedNumberOfMatches = 2
-  const matchIndexToReturn = 1
-  const taskRunIdRegex = /\sid(?:\s*=?\s*)([^\s)]*)(?:\s*\)?)/i
-  const taskRunIdText = 'task run ID'
-
-  const extractionData = {
-    taskRunData,
-    regex: taskRunIdRegex,
-    expectedNumberOfMatches,
-    matchIndexToReturn,
-    errorMessageSubject: taskRunIdText
-  }
-
-  return extract(context, extractionData)
+  return await extract(context, taskRunData, taskRunIdRegex, taskRunIdText)
 }
