+++ ProcessFewsEventCode/helpers/get-workflow-id.js
@@ -1,17 +1,8 @@
-const extract = require('./extract')
+const extract = require('./extraction-utils')
 
-const expectedNumberOfMatches = 2
-const matchIndexToReturn = 1
 const workflowIdRegex = /task(?:\s+run)?\s+([^\s]*)\s+/i
 const workflowIdText = 'workflow ID'
 
 module.exports = async function (context, taskRunData) {
-  const extractionData = {
-    taskRunData,
-    regex: workflowIdRegex,
-    expectedNumberOfMatches,
-    matchIndexToReturn,
-    errorMessageSubject: workflowIdText
-  }
-  return extract(context, extractionData)
+  return await extract(context, taskRunData, workflowIdRegex, workflowIdText)
 }
